# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sirene Automatique API Documentation",
        "description": "API Documentation for the Sirene Automatique project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/request-otp": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Request OTP for login",
                "description": "Demander un code OTP pour connexion",
                "operationId": "10dbaf9fdc7345c44ec7c44e73d10b35",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RequestOtpRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Code OTP envoyé avec succès."
                                        },
                                        "expires_in": {
                                            "type": "string",
                                            "example": "5 minutes"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "telephone": [
                                                    "Aucun compte associé à ce numéro de téléphone."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/verify-otp": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Verify OTP and log in",
                "description": "Vérifier l'OTP et se connecter",
                "operationId": "abcb5741589707b15c7ecac7a67c13aa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifyOtpRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Connexion réussie."
                                        },
                                        "access_token": {
                                            "description": "Bearer token",
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/AuthUser"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "otp": [
                                                    "Code OTP invalide ou expiré."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Log in with identifier and password",
                "description": "Connexion classique avec identifiant et mot de passe",
                "operationId": "dd51a484b7fa0f97c7680fdbde14cf5f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Connexion réussie."
                                        },
                                        "access_token": {
                                            "description": "Bearer token",
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/AuthUser"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "identifiant": [
                                                    "Identifiants incorrects."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Log out the authenticated user",
                "description": "Déconnexion",
                "operationId": "df9bdc6c7a6d9f20b983526b2ec0f698",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Déconnexion réussie."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get authenticated user details",
                "description": "Obtenir les informations de l'utilisateur connecté",
                "operationId": "b5de204f5e452ef51327c691aa99e77c",
                "responses": {
                    "200": {
                        "description": "User details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/AuthUser"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/auth/change-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Change authenticated user's password",
                "description": "Changer le mot de passe de l'utilisateur connecté",
                "operationId": "6ddeabdbd0e16e0dc7cda8677c0ba08e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangerMotDePasseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password changed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Mot de passe changé avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "ancien_mot_de_passe": [
                                                    "L'ancien mot de passe est incorrect."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AuthUser": {
                "title": "Authenticated User",
                "description": "User object returned upon successful authentication",
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "string",
                        "format": "uuid"
                    },
                    "nom_utilisateur": {
                        "description": "Username",
                        "type": "string"
                    },
                    "type": {
                        "description": "User type (e.g., ADMIN, ECOLE)",
                        "type": "string"
                    },
                    "telephone": {
                        "description": "User's phone number",
                        "type": "string",
                        "nullable": true
                    },
                    "email": {
                        "description": "User's email address",
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    },
                    "doit_changer_mot_de_passe": {
                        "description": "Indicates if user must change password",
                        "type": "boolean"
                    },
                    "mot_de_passe_change": {
                        "description": "Indicates if user has changed password at least once",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "ChangerMotDePasseRequest": {
                "title": "Change Password Request",
                "description": "Request body for changing user password",
                "required": [
                    "nouveau_mot_de_passe",
                    "nouveau_mot_de_passe_confirmation"
                ],
                "properties": {
                    "ancien_mot_de_passe": {
                        "description": "User's current password (optional, if admin is changing password)",
                        "type": "string",
                        "format": "password",
                        "nullable": true
                    },
                    "nouveau_mot_de_passe": {
                        "description": "User's new password",
                        "type": "string",
                        "format": "password",
                        "minLength": 8
                    },
                    "nouveau_mot_de_passe_confirmation": {
                        "description": "Confirmation of the new password",
                        "type": "string",
                        "format": "password",
                        "minLength": 8
                    }
                },
                "type": "object"
            },
            "LoginRequest": {
                "title": "Login Request",
                "description": "Request body for user login with identifier and password",
                "required": [
                    "identifiant",
                    "mot_de_passe"
                ],
                "properties": {
                    "identifiant": {
                        "description": "User's identifier (e.g., username, email, phone number)",
                        "type": "string",
                        "maxLength": 100
                    },
                    "mot_de_passe": {
                        "description": "User's password",
                        "type": "string",
                        "format": "password",
                        "minLength": 6
                    }
                },
                "type": "object"
            },
            "RequestOtpRequest": {
                "title": "Request OTP Request",
                "description": "Request body for requesting an OTP",
                "required": [
                    "telephone"
                ],
                "properties": {
                    "telephone": {
                        "description": "User's phone number",
                        "type": "string",
                        "pattern": "^[0-9]{8,15}$"
                    }
                },
                "type": "object"
            },
            "VerifyOtpRequest": {
                "title": "Verify OTP Request",
                "description": "Request body for verifying an OTP and logging in",
                "required": [
                    "telephone",
                    "otp"
                ],
                "properties": {
                    "telephone": {
                        "description": "User's phone number",
                        "type": "string",
                        "pattern": "^[0-9]{8,15}$"
                    },
                    "otp": {
                        "description": "OTP code",
                        "type": "string",
                        "maxLength": 6,
                        "minLength": 6,
                        "pattern": "^[0-9]{6}$"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}